+++ src/components/Nav.js
@@ -1,11 +1,6 @@
 import { Link } from 'react-router-dom';
 
 function Navbar() {
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        console.log(e.target.title.value);
-        e.target.reset()
-    }
     return (
         <div className="nav-container">
             <div className="navbar">
@@ -15,10 +10,6 @@ function Navbar() {
                     <Link to="/country" className="CATEGORIES">Country</Link>
                 </nav>
             </div>
-            <form action="#" onSubmit={(e)=>handleSubmit(e)} >
-                <input type="text" name='title' />
-                <button type="submit"> Search </button>
-            </form>
         </div>
     );
 }
